+++ src/embedded/mod.rs
@@ -18,6 +18,7 @@ pub mod generator;
 pub mod html;
 pub mod label;
 pub mod remove_reviewed;
+pub mod run_using;
 pub mod sink;
 
 #[derive(Debug)]